Setup lib and test. Rename directories with answers
